Move Mattermost commands into separate files
