+++ lib/commands.rb
@@ -41,133 +41,3 @@ class Commands
     )
   end
 end
-
-class Commands::WhereIs < Commands
-  def command
-    "WhereIs"
-  end
-
-  def response_body
-    if target_user.nil?
-      {
-        text: %(I couldn't find "#{target_username}".  Try typing their Mattermost name.)
-      }
-    elsif target_user.last_whereabouts.present?
-      time = target_user.last_whereabouts.sent_at
-      msg = "At #{time.strftime('%-l:%M%P')} on #{time.strftime('%m/%d/%y')}, #{target_user.name} sent \"#{target_user.last_whereabouts.body}\""
-      {
-        attachments: [{
-          fallback: msg,
-          title: "#{time.strftime('%-l:%M%P')}, #{time.strftime('%m/%d/%y')}:",
-          author_name: target_user.name,
-          text: target_user.last_whereabouts.body,
-          color: "#008800"
-        }]
-      }
-    else
-      { text: "#{target_user.name} hasn't set a where recently." }
-    end
-  end
-
-  def target_user
-    @user ||= User.find_by(email: "#{target_username}@eff.org")
-  end
-
-  private
-
-  def target_username
-    @body.split(' ').first.remove("@")
-  end
-end
-
-class Commands::SetMyWhere < Commands
-  def target_user
-    @user ||= User.find_by(email: "#{@username}@eff.org")
-  end
-
-  def command
-    "Where"
-  end
-
-  def response_body
-    { text: message }
-  end
-
-  def message
-    if target_user.present? && create_where
-      %(Your whereabouts are now set to "#{@body}".)
-    else
-      "I couldn't save your message. Better send it to where@eff.org :sweat_smile:"
-    end
-  end
-
-  private
-
-  def create_where
-    target_user.where_messages.create(
-      provenance: WhereMessage::PROVENANCES[:mattermost],
-      body: @body,
-      sent_at: DateTime.now
-    ).persisted?
-  end
-end
-
-class Commands::AddFabForward < Commands
-  def command
-    "AddFabForward"
-  end
-
-  def target_user
-    @user ||= User.find_by(email: "#{@username}@eff.org")
-  end
-
-  def response_body
-    { text: message }
-  end
-
-  def message
-    if target_user.present? && create_fab
-      %(You've added "#{@body}" to next week's FAB.)
-    else
-      "I couldn't save your FAB. Better set it at https://fab.eff.org/ :sweat_smile:"
-    end
-  end
-
-  def create_fab
-    fab = target_user.fabs.find_or_create_by(period: Fab.get_start_of_current_fab_period)
-    return false unless fab.persisted?
-
-    note = fab.forward.find { |note| note.body.nil? } || fab.forward.new
-    note.update(body: @body)
-  end
-end
-
-class Commands::AddFabBack < Commands
-  def command
-    "AddFabBack"
-  end
-
-  def target_user
-    @user ||= User.find_by(email: "#{@username}@eff.org")
-  end
-
-  def response_body
-    { text: message }
-  end
-
-  def message
-    if target_user.present? && create_fab
-      %(You've added "#{@body}" to this week's FAB.)
-    else
-      "I couldn't save your FAB. Better set it at https://fab.eff.org/ :sweat_smile:"
-    end
-  end
-
-  def create_fab
-    fab = target_user.fabs.find_or_create_by(period: Fab.get_start_of_current_fab_period)
-    return false unless fab.persisted?
-
-    note = fab.backward.find { |note| note.body.nil? } || fab.backward.new
-    note.update(body: @body)
-  end
-end
